Blast checkin of 53 files representing the first working 64-bit RTCP with TopologyGraph
git-svn-id: http://scm.sipfoundry.org/rep/sipX/branches/rtcp-reenable@12019 a612230a-c5fa-0310-af8b-88eea846685b

+++ sipXmediaLib/include/rtcp/ISetReceiverStatistics.h
@@ -1,4 +1,6 @@
 //
+// Copyright (C) 2006-2013 SIPez LLC.  All rights reserved.
+//
 // Copyright (C) 2004-2006 SIPfoundry Inc.
 // Licensed by SIPfoundry under the LGPL license.
 //
@@ -83,7 +85,7 @@ public:
  *
  *
  */
-    virtual void SetLastRcvdSRTime(unsigned long aulNTPTimestamp[]) = 0;
+    virtual void SetLastRcvdSRTime(uint32_t aulNTPTimestamp[]) = 0;
 
 };
